[PDF][MBL-14764] Part 1 of PSPDFKit api updates, toolbar items
